+++ src/rsg-components/Markdown/MarkdownTable/index.js
@@ -1,0 +1,5 @@
+export { default as MarkdownTable } from './MarkdownTableRenderer';
+export { default as TableHead } from './TableHeadRenderer';
+export { default as TableBody } from './TableBodyRenderer';
+export { default as TableRow } from './TableRowRenderer';
+export { default as TableCell } from './TableCellRenderer';