+++ src/rsg-components/Sections/SectionsRenderer.js
@@ -1,0 +1,2 @@
+import Section from 'bootstrap-styled/lib/Section';
+export default Section;