+++ styleguide.config.js
@@ -1,2 +1,3 @@
 const { config } = require('./lib/styleguide.config.js');
+
 module.exports = config;